internal/assettypes: rename assettype package to assettypes
